(PC-17814) identityCheck: move utils to /helpers
