+++ src/features/identityCheck/pages/phoneValidation/helpers/formatPhoneNumber.ts
